+++ src/main/java/com/finanzas/finanzasTF/repository/NegocioRepository.java
@@ -11,7 +11,7 @@ import java.util.List;
 
 @Repository
 public interface NegocioRepository extends JpaRepository<Negocio,Integer > {
-    Negocio findOneNegocioByRUC(Integer RUC);
+    Negocio findOneNegocioByRUC(Long RUC);
     Negocio findOneNegocioByPerfil_id(Integer perfil_id);
